requirements.txt
google-cloud-texttospeech>=2.19.0
requests>=2.28.0
